refactor(types): split mod.rs into oid.rs + traits.rs
Move Oid struct and constants to types/oid.rs, move ToSql/FromSql
traits and helpers to types/traits.rs. Re-export from mod.rs for
backwards compatibility.
